Bump version/refactor
