+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+0.2.5 (2025-06-14)
+==================
+* feat: Add support to view text-enabled plugins
+
 0.2.0 (2024-12-12)
 ==================
