+++ djangocms_text_ckeditor5/__init__.py
@@ -1,7 +1,7 @@
 from djangocms_text.editors import RTEConfig
 
 
-__version__ = "0.2.0"
+__version__ = "0.2.5"
 
 
 ckeditor5 = RTEConfig(
