+++ package.json
@@ -33,40 +33,41 @@
     "source-map-loader": "^5.0.0",
     "style-loader": "^4.0.0",
     "webpack": "^5.97.1",
-    "webpack-cli": "^5.1.4"
+    "webpack-cli": "^6.0.1"
   },
   "dependencies": {
-    "@ckeditor/ckeditor5-adapter-ckfinder": "^44.0.0",
-    "@ckeditor/ckeditor5-alignment": "^44.0.0",
-    "@ckeditor/ckeditor5-autoformat": "^44.0.0",
-    "@ckeditor/ckeditor5-autosave": "^44.0.0",
-    "@ckeditor/ckeditor5-basic-styles": "^44.0.0",
-    "@ckeditor/ckeditor5-block-quote": "^44.0.0",
-    "@ckeditor/ckeditor5-ckfinder": "^44.0.0",
-    "@ckeditor/ckeditor5-cloud-services": "^44.0.0",
-    "@ckeditor/ckeditor5-code-block": "^44.0.0",
+    "@ckeditor/ckeditor5-adapter-ckfinder": "^45.2.0",
+    "@ckeditor/ckeditor5-alignment": "^45.2.0",
+    "@ckeditor/ckeditor5-autoformat": "^45.2.0",
+    "@ckeditor/ckeditor5-autosave": "^45.2.0",
+    "@ckeditor/ckeditor5-basic-styles": "^45.2.0",
+    "@ckeditor/ckeditor5-block-quote": "^45.2.0",
+    "@ckeditor/ckeditor5-ckfinder": "^45.2.0",
+    "@ckeditor/ckeditor5-cloud-services": "^45.2.0",
+    "@ckeditor/ckeditor5-code-block": "^45.2.0",
     "@ckeditor/ckeditor5-dev-utils": "^43.0.1",
-    "@ckeditor/ckeditor5-easy-image": "^44.0.0",
-    "@ckeditor/ckeditor5-editor-balloon": "^44.0.0",
-    "@ckeditor/ckeditor5-editor-classic": "^44.0.0",
-    "@ckeditor/ckeditor5-editor-inline": "^44.0.0",
-    "@ckeditor/ckeditor5-essentials": "^44.0.0",
-    "@ckeditor/ckeditor5-font": "^44.0.0",
-    "@ckeditor/ckeditor5-heading": "^44.0.0",
-    "@ckeditor/ckeditor5-highlight": "^44.0.0",
-    "@ckeditor/ckeditor5-horizontal-line": "^44.0.0",
-    "@ckeditor/ckeditor5-image": "^44.0.0",
-    "@ckeditor/ckeditor5-indent": "^44.0.0",
-    "@ckeditor/ckeditor5-link": "^44.0.0",
-    "@ckeditor/ckeditor5-list": "^44.0.0",
-    "@ckeditor/ckeditor5-media-embed": "^44.0.0",
-    "@ckeditor/ckeditor5-paragraph": "^44.0.0",
-    "@ckeditor/ckeditor5-paste-from-office": "^44.0.0",
-    "@ckeditor/ckeditor5-source-editing": "^44.0.0",
-    "@ckeditor/ckeditor5-table": "^44.0.0",
-    "@ckeditor/ckeditor5-theme-lark": "^44.0.0",
-    "@ckeditor/ckeditor5-typing": "^44.0.0",
-    "@ckeditor/ckeditor5-upload": "^44.0.0",
+    "@ckeditor/ckeditor5-easy-image": "^45.2.0",
+    "@ckeditor/ckeditor5-editor-balloon": "^45.2.0",
+    "@ckeditor/ckeditor5-editor-classic": "^45.2.0",
+    "@ckeditor/ckeditor5-editor-inline": "^45.2.0",
+    "@ckeditor/ckeditor5-essentials": "^45.2.0",
+    "@ckeditor/ckeditor5-font": "^45.2.0",
+    "@ckeditor/ckeditor5-heading": "^45.2.0",
+    "@ckeditor/ckeditor5-highlight": "^45.2.0",
+    "@ckeditor/ckeditor5-horizontal-line": "^45.2.0",
+    "@ckeditor/ckeditor5-icons": "^45.2.0",
+    "@ckeditor/ckeditor5-image": "^45.2.0",
+    "@ckeditor/ckeditor5-indent": "^45.2.0",
+    "@ckeditor/ckeditor5-link": "^45.2.0",
+    "@ckeditor/ckeditor5-list": "^45.2.0",
+    "@ckeditor/ckeditor5-media-embed": "^45.2.0",
+    "@ckeditor/ckeditor5-paragraph": "^45.2.0",
+    "@ckeditor/ckeditor5-paste-from-office": "^45.2.0",
+    "@ckeditor/ckeditor5-source-editing": "^45.2.0",
+    "@ckeditor/ckeditor5-table": "^45.2.0",
+    "@ckeditor/ckeditor5-theme-lark": "^45.2.0",
+    "@ckeditor/ckeditor5-typing": "^45.2.0",
+    "@ckeditor/ckeditor5-upload": "^45.2.0",
     "html-loader": "^5.1.0"
   }
 }
