+++ private/js/ckeditor5_plugins/ckeditor5.cmslink/index.js
