+++ private/js/ckeditor5_plugins/ckeditor5.cmsplugin/src/cmsplugin.js
