+++ private/js/ckeditor5_plugins/ckeditor5.cmsplugin/src/cmspluginui.js
@@ -10,7 +10,7 @@ import { Collection } from 'ckeditor5/src/utils';
 import { Model, createDropdown, addListToDropdown } from 'ckeditor5/src/ui';
 import ButtonView from '@ckeditor/ckeditor5-ui/src/button/buttonview';
 
-import cmsPluginIcon from './../theme/icons/puzzle.svg';
+import cmsPluginIcon from '../theme/icons/puzzle.svg';
 import { Modal } from './modal/modalui';
 import {addLinkProtocolIfApplicable} from "@ckeditor/ckeditor5-link/src/utils";
