+++ webpack.config.js
@@ -30,7 +30,7 @@ module.exports = {
                 use: ['css-loader'],
             },
             {
-                test: /ckeditor5-[^/\\]+[/\\]theme[/\\]icons[/\\][^/\\]+\.svg$/,
+                test: /ckeditor5[^/\\]+[/\\]theme[/\\]icons[/\\][^/\\]+\.svg$/,
                 use: ['raw-loader']
             },
             {
